// src/features/KumikoEditor/models/Settings.ts
/** Manages the Project-specific settings */
export interface Settings {
  /** width of the panel */
  width: number;
 /** height of the panel */
  height: number;
 /** depth of the panel */
  depth: number;
 /** framethickness */
  frameWidth: number;
 /** width of single wood strips*/
  lineWidth: number;
 /** spacing based on the ground side of each triangle */
  spacing: number;
}